@@ -8,21 +8,22 @@
     "ts-node": "ts-node",
     "start": "ts-node index.ts",
     "dev": "ts-node-dev index.ts",
-    "lint": "eslint .",
+    "lint": "eslint --ext .ts .",
     "migration:down": "ts-node utils/rollMigrations.ts",
     "migration:up": "ts-node --transpile-only utils/runMigrations.ts"
   },
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@types/cors": "^2.8.13",
     "@types/express": "^4.17.17",
-    "@types/node": "^18.14.6",
+    "@types/node": "^18.15.3",
     "@types/pg": "^8.6.6",
     "@types/sequelize": "^4.28.14",
     "@types/validator": "^13.7.13",
-    "@typescript-eslint/eslint-plugin": "^5.54.0",
-    "@typescript-eslint/parser": "^5.54.0",
-    "eslint": "^8.35.0",
+    "@typescript-eslint/eslint-plugin": "^5.55.0",
+    "@typescript-eslint/parser": "^5.55.0",
+    "eslint": "^8.36.0",
     "esm": "^3.2.25",
     "pg": "^8.10.0",
     "pg-hstore": "^2.3.4",
